# src/main/resources/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "contact" : {
      "name" : "PowerCards Support",
      "url" : "https://github.com/victobonetti/powercards"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "PowerCards API",
    "version" : "1.0.0"
  },
  "paths" : {
...
    "/v1/notes" : {
      "get" : {
        "summary" : "List all notes",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1
          }
        }, {
          "name" : "perPage",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        }, {
          "name" : "search",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaginatedResponseNoteResponse"
                }
              }
            }
          }
        },
        "tags" : [ "Note Resource" ]
      },
      "post" : {
        "summary" : "Create a new note",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NoteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Note created"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "tags" : [ "Note Resource" ]
      }
    },
    "/v1/notes/bulk/delete" : {
      "post" : {
        "summary" : "Bulk delete notes",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BulkDeleteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "tags" : [ "Note Resource" ]
      }
    },
    "/v1/notes/bulk/tags" : {
      "post" : {
        "summary" : "Bulk add tags to notes",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BulkTagRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "tags" : [ "Note Resource" ]
      }
    },
    "/v1/notes/{id}" : {
      "put" : {
        "summary" : "Update an existing note",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NoteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Note updated",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/NoteResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Note not found"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "tags" : [ "Note Resource" ]
      },
      "get" : {
        "summary" : "Get a note by ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Note found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/NoteResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Note not found"
          }
        },
        "tags" : [ "Note Resource" ]
      },
      "delete" : {
        "summary" : "Delete a note",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Note deleted"
          },
          "404" : {
            "description" : "Note not found"
          }
        },
        "tags" : [ "Note Resource" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "BulkDeleteRequest" : {
        "type" : "object",
        "properties" : {
          "ids" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        }
      },
...
      "BulkTagRequest" : {
        "type" : "object",
        "properties" : {
          "noteIds" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "NoteRequest" : {
        "type" : "object",
        "properties" : {
          "modelId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "modificationTimestamp" : {
            "type" : "integer",
            "format" : "int64"
          },
          "updateSequenceNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "tags" : {
            "type" : "string"
          },
          "fields" : {
            "type" : "string"
          },
          "sortField" : {
            "type" : "string"
          },
          "checksum" : {
            "type" : "integer",
            "format" : "int64"
          },
          "flags" : {
            "type" : "integer",
            "format" : "int32"
          },
          "customData" : {
            "type" : "string"
          }
        }
      },
      "NoteResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "guid" : {
            "type" : "string"
          },
          "modelId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "modificationTimestamp" : {
            "type" : "integer",
            "format" : "int64"
          },
          "updateSequenceNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "tags" : {
            "type" : "string"
          },
          "fields" : {
            "type" : "string"
          },
          "sortField" : {
            "type" : "string"
          },
          "checksum" : {
            "type" : "integer",
            "format" : "int64"
          },
          "flags" : {
            "type" : "integer",
            "format" : "int32"
          },
          "customData" : {
            "type" : "string"
          }
        }
      },
...
      "PaginatedResponseNoteResponse" : {
        "type" : "object",
        "properties" : {
          "pagination" : {
            "$ref" : "#/components/schemas/PaginationMeta"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/NoteResponse"
            }
          }
        }
      },
...
      "PaginationMeta" : {
        "type" : "object",
        "properties" : {
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "nextPageUri" : {
            "type" : "string"
          },
          "lastPageUri" : {
            "type" : "string"
          }
        }
      },
...
    }
  },
  "servers" : [ {
    "url" : "http://localhost:8088",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8088",
    "description" : "Auto generated value"
  } ]
}